chat/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.4
chardet==4.0.0
click==8.1.7
cryptography==3.4.8
distro==1.7.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
httplib2==0.20.2
httptools==0.6.2
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.3.0
pyparsing==2.4.7
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==5.4.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
typing_extensions==4.12.2
uvicorn==0.31.1
uvloop==0.21.0
wadllib==1.3.6
watchfiles==0.24.0
websockets==13.1
zipp==1.0.0